backend/text_to_video/requirements.txt
moviepy==1.0.3
# fastapi==0.104.1
uvicorn==0.23.2
# pydantic==2.10.6
replicate==1.0.4
python-dotenv==1.0.1
groq==0.18.0
# instructor==1.7.2
pydub==0.25.1
anthropic==0.49.0
captacity==0.3.1
openai-whisper
# aiohttp==3.11.13
# rich==13.9.4
# typer==0.9.4
elevenlabs==1.52.0
# Added for HeyGen Workflow
requests # For API calls
boto3 # For AWS S3 interaction
